# docs/swagger.yaml
definitions:
...
  models.ErrorResponse:
    properties:
      code:
        description: code
        type: integer
      details:
        description: details
        type: string
      fields:
        description: fields
        items:
          type: string
        type: array
      message:
        description: message
        type: string
      result:
        description: result
        type: string
      sub-code:
        description: sub code
        type: integer
    type: object
...
  models.InstanceSnapshot:
    properties:
      checksum:
        description: '"sha256:<hex>", gateway-computed (envelope)'
        type: string
      checksum_ok:
        description: integrity-sweep verdict
        type: boolean
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      encrypted:
        type: boolean
      gateway_version:
        type: string
      id:
        type: string
      instance_id:
        type: integer
      last_restore_response:
        description: |-
          LastRestoreResponse is the full gateway response JSON of the most
          recent restore attempt (the audit record). Only populated on the
          single-snapshot GET, not in lists.
        type: string
      last_restore_result:
        type: string
      last_restored_at:
        type: string
      name:
        type: string
      pinned:
        type: boolean
      restore_count:
        type: integer
      schema_version:
        type: string
      size_bytes:
        description: uncompressed JSON size
        type: integer
      stored_checksum:
        description: '"sha256:<hex>" over raw JSON bytes as received, OAM-computed'
        type: string
      trigger_type:
        type: string
    type: object
  models.InstanceSnapshotSchedule:
    properties:
      enabled:
        type: boolean
      instance_id:
        type: integer
      interval_hours:
        type: integer
      last_run_at:
        type: string
      last_run_result:
        type: string
      retain_count:
        type: integer
    type: object
...
  models.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  models.PaginatedSnapshotsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.InstanceSnapshot'
        type: array
      pagination:
        $ref: '#/definitions/models.PaginationMeta'
    type: object
  models.PaginationMeta:
    properties:
      has_next:
        description: Whether there's a next page
        type: boolean
      has_prev:
        description: Whether there's a previous page
        type: boolean
      limit:
        description: Number of items per page
        type: integer
      page:
        description: Current page number
        type: integer
      total_count:
        description: Total number of items
        type: integer
      total_pages:
        description: Total number of pages
        type: integer
    type: object
  models.RestoreOutcome:
    properties:
      cross_instance:
        type: boolean
      gateway_response:
        type: object
      gateway_status:
        type: integer
      instance_id:
        description: restore target
        type: integer
      mode:
        type: string
      pre_restore_snapshot_id:
        type: string
      snapshot_id:
        type: string
    type: object
  models.RestoreSnapshotRequest:
    properties:
      mode:
        description: '"dry-run" (default) | "commit"'
        type: string
      target_instance_id:
        description: |-
          TargetInstanceID restores the snapshot onto a different instance than
          the one it was taken from (cross-instance restore). Defaults to the
          snapshot's own instance.
        type: integer
    type: object
...
  models.SnapshotScheduleRequest:
    properties:
      enabled:
        type: boolean
      interval_hours:
        maximum: 8760
        minimum: 1
        type: integer
      retain_count:
        maximum: 1000
        minimum: 1
        type: integer
    type: object
...
  models.TakeSnapshotRequest:
    properties:
      description:
        type: string
      name:
        type: string
      trigger_type:
        description: defaults to "manual"
        type: string
    type: object
...
  models.UpdateSnapshotRequest:
    properties:
      description:
        type: string
      name:
        type: string
      pinned:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /oam/instances/{id}/snapshot-schedule:
    get:
      description: Returns the scheduled-snapshot/retention settings; defaults (disabled,
        every 24h, keep 10) when never configured.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: LoxiLB Instance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.InstanceSnapshotSchedule'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Read an instance's snapshot schedule
      tags:
      - snapshots
    put:
      consumes:
      - application/json
      description: Enables/disables scheduled snapshots and sets interval and per-instance
        retention (keep-N unpinned; pre_upgrade and pinned snapshots are exempt).
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: LoxiLB Instance ID
        in: path
        name: id
        required: true
        type: integer
      - description: Schedule settings
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SnapshotScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.InstanceSnapshotSchedule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an instance's snapshot schedule
      tags:
      - snapshots
  /oam/instances/{id}/snapshots:
    get:
      description: Returns snapshot metadata (never blobs) for one instance, newest
        first, paginated.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: LoxiLB Instance ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedSnapshotsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List snapshots of an instance
      tags:
      - snapshots
    post:
      consumes:
      - application/json
      description: Calls the gateway's GET /config/snapshot on the managed instance
        and stores the document (gzip, AES-256-GCM at rest when SNAPSHOT_ENC_KEY is
        set) in the OAM database. Returns metadata only, never the blob.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: LoxiLB Instance ID
        in: path
        name: id
        required: true
        type: integer
      - description: Snapshot name/description/trigger
        in: body
        name: request
        schema:
          $ref: '#/definitions/models.TakeSnapshotRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.InstanceSnapshot'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "502":
          description: Gateway unreachable (connection error passed through verbatim)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Take an instance config snapshot now
      tags:
      - snapshots
  /oam/instances/{id}/snapshots/upload:
    post:
      consumes:
      - multipart/form-data
      description: Accepts a previously downloaded snapshot document (multipart field
        "file"). Only the envelope (schema_version, gateway_version, checksum) is
        parsed — deep validation stays the gateway's job at restore time.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: LoxiLB Instance ID
        in: path
        name: id
        required: true
        type: integer
      - description: Snapshot document JSON
        in: formData
        name: file
        required: true
        type: file
      - description: Snapshot name
        in: formData
        name: name
        type: string
      - description: Description
        in: formData
        name: description
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.InstanceSnapshot'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Re-import an off-box snapshot archive
      tags:
      - snapshots
...
  /oam/snapshots/{sid}:
    delete:
      description: Deletes a stored snapshot. Pinned snapshots require force=true.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Snapshot ID (UUID)
        in: path
        name: sid
        required: true
        type: string
      - description: Required to delete a pinned snapshot
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Snapshot is pinned and force was not set
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a snapshot
      tags:
      - snapshots
    get:
      description: Returns snapshot metadata including restore history and the full
        gateway response of the last restore (the audit record).
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Snapshot ID (UUID)
        in: path
        name: sid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.InstanceSnapshot'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get one snapshot's metadata
      tags:
      - snapshots
    patch:
      consumes:
      - application/json
      description: Updates name, description and/or pinned. Pinned snapshots are exempt
        from retention.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Snapshot ID (UUID)
        in: path
        name: sid
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateSnapshotRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.InstanceSnapshot'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update snapshot metadata
      tags:
      - snapshots
  /oam/snapshots/{sid}/download:
    get:
      description: Streams the decrypted, decompressed snapshot JSON. The document
        contains IPsec PSKs and certificate private keys, so this is write-gated and
        audit-logged. X-Snapshot-Checksum carries the gateway's document checksum;
        X-Content-Checksum is sha256 over the exact bytes served.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Snapshot ID (UUID)
        in: path
        name: sid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: snapshot document JSON
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "422":
          description: Stored blob failed integrity verification
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Download a snapshot document
      tags:
      - snapshots
  /oam/snapshots/{sid}/restore:
    post:
      consumes:
      - application/json
      description: 'Default mode is dry-run: the gateway validates and returns its
        plan without mutating anything. Commit first takes an automatic pre_restore
        safety snapshot of the target, then applies. The gateway''s response is returned
        verbatim in gateway_response. Cross-instance restore is allowed and flagged
        with cross_instance=true.'
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Snapshot ID (UUID)
        in: path
        name: sid
        required: true
        type: string
      - description: 'mode: dry-run (default) | commit; optional target_instance_id'
        in: body
        name: request
        schema:
          $ref: '#/definitions/models.RestoreSnapshotRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RestoreOutcome'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "422":
          description: Stored blob failed integrity verification (never sent to the
            gateway)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "502":
          description: Gateway unreachable (connection error passed through verbatim)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Restore a stored snapshot to a gateway
      tags:
      - snapshots
...
securityDefinitions:
  BearerAuth:
    description: JWT bearer token from POST /oam/login, sent as "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"